requirements.txt
python==3.9.13
Django
pillow
django-filter
